+++ air/tests/test_module/features/mod.rs
@@ -19,5 +19,6 @@ mod errors;
 mod join_behaviour;
 mod lambda;
 mod misc;
+mod scopes;
 mod streams;
 mod tetraplets;