+++ crates/air-lib/air-parser/src/parser/lexer/errors.rs
@@ -14,52 +14,140 @@
  * limitations under the License.
  */
 
+use crate::parser::Span;
 use thiserror::Error as ThisError;
 
 use std::num::ParseFloatError;
 use std::num::ParseIntError;
+use std::ops::Range;
 
 #[derive(ThisError, Debug, Clone, PartialEq, Eq)]
 pub enum LexerError {
     #[error("this string literal has unclosed quote")]
-    UnclosedQuote(usize, usize),
+    UnclosedQuote(Span),
 
     #[error("empty string aren't allowed in this position")]
-    EmptyString(usize, usize),
+    EmptyString(Span),
 
     #[error("only alphanumeric, '_', and '-' characters are allowed in this position")]
-    IsNotAlphanumeric(usize, usize),
+    IsNotAlphanumeric(Span),
 
     #[error("a stream name should be non empty")]
-    EmptyStreamName(usize, usize),
+    EmptyStreamName(Span),
 
     #[error("this variable or constant shouldn't have empty name")]
-    EmptyVariableOrConst(usize, usize),
+    EmptyVariableOrConst(Span),
 
     #[error("invalid character in lambda")]
-    InvalidLambda(usize, usize),
+    InvalidLambda(Span),
 
     #[error("a digit could contain only digits or one dot")]
-    UnallowedCharInNumber(usize, usize),
+    UnallowedCharInNumber(Span),
 
-    #[error("{2}")]
-    ParseIntError(usize, usize, #[source] ParseIntError),
+    #[error("{1}")]
+    ParseIntError(Span, #[source] ParseIntError),
 
-    #[error("{2}")]
-    ParseFloatError(usize, usize, #[source] ParseFloatError),
+    #[error("{1}")]
+    ParseFloatError(Span, #[source] ParseFloatError),
 
     // TODO: use LambdaParserError directly here (it'll require introducing a lifetime)
-    #[error("{2}")]
-    LambdaParserError(usize, usize, String),
+    #[error("{se_lambda_parser_error}")]
+    LambdaParserError {
+        span: Span,
+        se_lambda_parser_error: String,
+    },
 
-    #[error("{2} is an incorrect path for %last_error%, only .$.instruction, .$.msg, and .$.peer_id are allowed")]
-    LastErrorPathError(usize, usize, String),
+    #[error("{error_path} is an incorrect path for %last_error%, only .$.instruction, .$.msg, and .$.peer_id are allowed")]
+    LastErrorPathError { span: Span, error_path: String },
 
     #[error("this float is too big, a float could contain less than 12 digits")]
-    TooBigFloat(usize, usize),
+    TooBigFloat(Span),
 
     #[error("leading dot without any symbols before - please write 0 if it's float or variable name if it's a lambda")]
-    LeadingDot(usize, usize),
+    LeadingDot(Span),
+}
+
+impl LexerError {
+    pub fn span(&self) -> Span {
+        let span = match self {
+            Self::UnclosedQuote(span) => span,
+            Self::EmptyString(span) => span,
+            Self::IsNotAlphanumeric(span) => span,
+            Self::EmptyStreamName(span) => span,
+            Self::EmptyVariableOrConst(span) => span,
+            Self::InvalidLambda(span) => span,
+            Self::UnallowedCharInNumber(span) => span,
+            Self::ParseIntError(span, _) => span,
+            Self::ParseFloatError(span, _) => span,
+            Self::LambdaParserError { span, .. } => span,
+            Self::LastErrorPathError { span, .. } => span,
+            Self::TooBigFloat(span) => span,
+            Self::LeadingDot(span) => span,
+        };
+
+        *span
+    }
+
+    pub fn unclosed_quote(range: Range<usize>) -> Self {
+        Self::UnclosedQuote(range.into())
+    }
+
+    pub fn empty_string(range: Range<usize>) -> Self {
+        Self::EmptyString(range.into())
+    }
+
+    pub fn is_not_alphanumeric(range: Range<usize>) -> Self {
+        Self::IsNotAlphanumeric(range.into())
+    }
+
+    pub fn empty_stream_name(range: Range<usize>) -> Self {
+        Self::EmptyStreamName(range.into())
+    }
+
+    pub fn empty_variable_or_const(range: Range<usize>) -> Self {
+        Self::EmptyVariableOrConst(range.into())
+    }
+
+    pub fn invalid_lambda(range: Range<usize>) -> Self {
+        Self::InvalidLambda(range.into())
+    }
+
+    pub fn unallowed_char_in_number(range: Range<usize>) -> Self {
+        Self::UnallowedCharInNumber(range.into())
+    }
+
+    pub fn parse_int_error(range: Range<usize>, parse_int_error: ParseIntError) -> Self {
+        Self::ParseIntError(range.into(), parse_int_error)
+    }
+
+    pub fn parse_float_error(range: Range<usize>, parse_float_error: ParseFloatError) -> Self {
+        Self::ParseFloatError(range.into(), parse_float_error)
+    }
+
+    pub fn lambda_parser_error(
+        range: Range<usize>,
+        se_lambda_parser_error: impl Into<String>,
+    ) -> Self {
+        Self::LambdaParserError {
+            span: range.into(),
+            se_lambda_parser_error: se_lambda_parser_error.into(),
+        }
+    }
+
+    pub fn last_error_path_error(range: Range<usize>, error_path: String) -> Self {
+        Self::LastErrorPathError {
+            span: range.into(),
+            error_path,
+        }
+    }
+
+    pub fn too_big_float(range: Range<usize>) -> Self {
+        Self::TooBigFloat(range.into())
+    }
+
+    pub fn leading_dot(range: Range<usize>) -> Self {
+        Self::LeadingDot(range.into())
+    }
 }
 
 use super::Token;