+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Version 0.22.0 (2021-04-14)
+
+[PR 243](https://github.com/fluencelabs/aquavm/pull/243):
+Clean scalars at the end of scope, only one next in a fold over scalar is possible now
+
+[PR 231](https://github.com/fluencelabs/aquavm/pull/231):
+Test refactoring
+
+[PR 228](https://github.com/fluencelabs/aquavm/pull/228):
+Improve stream determinism
+
 ## Version 0.21.0 (2021-02-26)
 
 [PR 225](https://github.com/fluencelabs/aquavm/pull/225):